refactor(ChatPanel): mv quick reply components to own files
